requirements.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
boto3==1.23.10
botocore==1.26.10
certifi==2022.9.24
charset-normalizer==2.0.12
cssmin==0.2.0
dj-database-url==1.0.0
Django==3.2.10
django-appconf==1.0.5
django-compressor==4.3
django-embed-video==1.4.8
django-filter==21.1
django-heroku==0.3.1
django-htmlmin==0.11.0
django-makeallmessages==0.2.1
django-modeltranslation==0.18.5
django-rosetta==0.9.8
django-storages==1.12.3
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==4.8.3
jmespath==0.10.0
jsmin==3.0.1
Pillow==8.4.0
polib==1.1.1
psycopg2==2.9.4
python-dateutil==2.8.2
python-decouple==3.6
python-memcached==1.59
pytz==2021.3
rcssmin==1.1.1
requests==2.27.1
rjsmin==1.2.1
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.12
webencodings==0.5.1
whitenoise==5.3.0
zipp==3.6.0
